Removed sensitive log in Verify
If a user did not exist in the db, but Verify searches for it, the old
behavior was to log an error detailing that this had happened. However,
this information is potentially sensitive, so it's better to be vague
and show the same credential-mismatch error given by inputting a wrong
password.